server/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
Flask-Cors==3.0.9
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mtgsdk==1.3.1
pymongo==3.12.0
python-dotenv==0.14.0
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
urllib3==1.25.11
Werkzeug==1.0.1
